lms/requirements.txt
django==4.2.21
jupyter_client==8.6.3
nbclient==0.10.2
nbformat==5.10.4
pydantic==2.11.4
pyyaml==6.0.2
ipykernel==6.29.5
websocket-client==1.8.0
pytils==0.4.3
celery==5.5.2
redis==6.0.0
numpy==2.0.2
jupyterhub==5.3.0
jupyter_server>=2.0.0
jupyterlab>=4.0.0
memory_profiler==0.61.0